chore: move campaign parser to core
