+++ packages/analytics-browser/src/types.ts
@@ -36,6 +36,8 @@ export {
   DEFAULT_CSS_SELECTOR_ALLOWLIST,
   DEFAULT_DATA_ATTRIBUTE_PREFIX,
   DEFAULT_ACTION_CLICK_ALLOWLIST,
+  Campaign,
+  ClickIdParameters,
+  ReferrerParameters,
+  UTMParameters,
 } from '@amplitude/analytics-core';
-
-export { Campaign, CampaignParser, ClickIdParameters, ReferrerParameters, UTMParameters } from './attribution/campaign';
